requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=2.13.0
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
python-multipart==0.0.6
google-ads>=25.1.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
pandas==2.2.0
openpyxl==3.1.2
gspread>=6.0.0
pytz>=2024.1
# R-URL Optimizer
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
tqdm>=4.65.0
# The vendored SA360 SDK imports pkg_resources, which setuptools 81+ no longer
# ships by default. Pin to an older setuptools until Google ships an SDK build
# that drops the pkg_resources dependency.
setuptools<81
./backend/vendor/searchads360-py.tar.gz
